+++ test/instances/Value.hs
@@ -1,17 +1,16 @@
 module Value (properties) where
 
 import Helpers (failProperty, sortedUnique)
-import Plutarch.Test.QuickCheck.Instances ()
+import Plutarch.Test.QuickCheck.Modifiers (GenValue (GenValue))
 import PlutusLedgerApi.V2 (TokenName, Value (Value))
 import qualified PlutusTx.AssocMap as AssocMap
 import Test.QuickCheck (
   Arbitrary (arbitrary, shrink),
-  Property,
+  Property, NonNegative (NonNegative), Positive (Positive),
   conjoin,
   counterexample,
   discard,
   forAllShrinkShow,
-  property,
   (.&&.),
   (===),
  )
@@ -35,29 +34,30 @@ properties =
 -- - Non-empty
 -- - First entry must be the ADA symbol
 propCSKeys :: Property
-propCSKeys = forAllShrinkShow arbitrary shrink ppShow $ \(Value val) ->
-  case fst <$> AssocMap.toList val of
-    [] -> counterexample "Empty 'outer map'." failProperty
-    keys@(sym : _) -> sortedUnique keys .&&. (sym === "")
+propCSKeys = forAllShrinkShow arbitrary shrink ppShow $ 
+  \(GenValue @NonNegative @Positive (Value rep)) ->
+    case fst <$> AssocMap.toList rep of
+      [] -> counterexample "Empty 'outer map'." failProperty
+      keys@(sym : _) -> sortedUnique keys .&&. (sym === "")
 
--- The ADA entry should be a singleton map, with the ADA token name as a key,
--- and a non-negative value
+-- The ADA entry should be a singleton map, with the ADA token name as a key
 propAdaInner :: Property
-propAdaInner = forAllShrinkShow arbitrary shrink ppShow $ \(Value val) ->
-  case AssocMap.lookup "" val of
+propAdaInner = forAllShrinkShow arbitrary shrink ppShow $ 
+  \(GenValue @NonNegative @Positive (Value rep)) ->
+  case AssocMap.lookup "" rep of
     Nothing -> counterexample "ADA entry not found." failProperty
     Just inner -> case AssocMap.toList inner of
       [] -> counterexample "ADA entry has empty 'inner map'." failProperty
-      [(tn, amount)] -> (tn === "") .&&. property (amount >= 0)
+      [(tn, _)] -> tn === ""
       _ -> counterexample "ADA entry is not a singleton." failProperty
 
 -- A non-ADA entry should:
 -- - Be non-empty
 -- - Have sorted, unique keys
--- - Have positive values
 propOtherInner :: Property
-propOtherInner = forAllShrinkShow arbitrary shrink ppShow $ \(Value val) ->
-  case AssocMap.toList val of
+propOtherInner = forAllShrinkShow arbitrary shrink ppShow $ 
+  \(GenValue @NonNegative @Positive (Value rep)) ->
+  case AssocMap.toList rep of
     [] -> counterexample "Empty 'outer map'." failProperty
     [_] -> discard
     (_ : rest) ->
@@ -67,7 +67,4 @@ propOtherInner = forAllShrinkShow arbitrary shrink ppShow $ \(Value val) ->
     go :: [(TokenName, Integer)] -> Property
     go = \case
       [] -> counterexample "Empty 'inner map'." failProperty
-      inner ->
-        let keys = fst <$> inner
-            vals = snd <$> inner
-         in sortedUnique keys .&&. conjoin (property . (> 0) <$> vals)
+      inner -> sortedUnique $ fst <$> inner